refactor class files for organization
