+++ App/app/src/main/java/com/flaredown/flaredownApp/Models/Alarm.java
@@ -1,4 +1,4 @@
-package com.flaredown.flaredownApp.FlareDown;
+package com.flaredown.flaredownApp.Models;
 
 import java.util.Random;
