+++ App/app/src/main/java/com/flaredown/flaredownApp/Receivers/InternetStatusBroadcastReceiver.java
@@ -1,4 +1,4 @@
-package com.flaredown.flaredownApp;
+package com.flaredown.flaredownApp.Receivers;
 
 import android.content.BroadcastReceiver;
 import android.content.Context;
