+++ App/app/src/main/res/layout/activity_settings.xml
@@ -8,7 +8,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical"
-    tools:context="com.flaredown.flaredownApp.SettingsActivity">
+    tools:context=".Settings.SettingsActivity">
 
     <com.flaredown.flaredownApp.MainToolbarView
         android:id="@+id/main_toolbar_view"
