+++ App/app/src/main/res/menu/menu_policy_webview.xml
@@ -1,3 +1,3 @@
 <menu xmlns:tools="http://schemas.android.com/tools"
-    tools:context="com.flaredown.flaredownApp.PolicyWebView">
+    tools:context=".Settings.PolicyWebView">
 </menu>
