+++ App/app/src/main/res/menu/menu_settings.xml
@@ -1,7 +1,7 @@
 <menu xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
-    tools:context="com.flaredown.flaredownApp.SettingsActivity">
+    tools:context=".Settings.SettingsActivity">
     <item
         android:id="@+id/action_save"
         android:orderInCategory="100"
